LayoutTests/accessibility/mac/live-regions/live-region-languages.html

<!DOCTYPE HTML><!-- webkit-test-runner [ IsAriaLiveRegionManagementEnabled=true ] -->
<html>
<head>
<script src="../../../resources/js-test.js"></script>
<script src="../../../resources/accessibility-helper.js"></script>
<meta http-equiv="content-language" content="en">
</head>
<body>

<div id="region" aria-live="polite">
<span lang="en">Apple:</span>
<span lang="es" id="spanish-region"></span>
</div>

<script>
var output = "This tests that live region notifications for elements with different languages include the right language attributes.\n\n";    
var testFinished = false;

function notificationCallback(object, notification, userInfo) {
    if (notification == "AXAnnouncementRequested") {
        output += `Received announcement request:\n${formatAnnouncementUserInfo(userInfo)}\n`;

        testFinished = true;
    }	
}

if (accessibilityController) {
    window.jsTestIsAsync = true;
    accessibilityController.addNotificationListener(notificationCallback);
    accessibilityController.accessibleElementById("polite-region");

    document.getElementById("spanish-region").textContent = "Manzana";
 
    setTimeout(async function() {
        await waitFor(() => testFinished);
        accessibilityController.removeNotificationListener();
        debug(output);
        finishJSTest();
    }, 0);
}
</script>
</body>
</html>
